+++ app/build.gradle
@@ -74,4 +74,6 @@ dependencies {
     implementation 'androidx.activity:activity-ktx:1.1.0'
     // Coil
     implementation("io.coil-kt:coil:2.5.0")
+    // coroutine
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.9'
 }
