+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/home/FriendProfileViewHolder.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.data.adapter.home
+package org.sopt.dosopttemplate.presentation.main.home
 
 import androidx.recyclerview.widget.RecyclerView
 import org.sopt.dosopttemplate.databinding.ItemFriendBinding
