+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/home/FriendSealed.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.data.adapter.home
+package org.sopt.dosopttemplate.presentation.main.home
 
 sealed class FriendSealed {
