+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/home/MyProfileViewHolder.kt
@@ -1,7 +1,7 @@
-package org.sopt.dosopttemplate.data.adapter.home
+package org.sopt.dosopttemplate.presentation.main.home
 
 import androidx.recyclerview.widget.RecyclerView
-import org.sopt.dosopttemplate.data.adapter.home.FriendSealed
+import org.sopt.dosopttemplate.presentation.main.home.FriendSealed
 import org.sopt.dosopttemplate.databinding.ItemMyProfileBinding
 
 class MyProfileViewHolder(private val binding: ItemMyProfileBinding) : RecyclerView.ViewHolder(binding.root) {
